Backend/requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
chess==1.11.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
python-dotenv
python-multipart
opencv-python
Pillow
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mypy==1.16.0
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-chess==1.999
requests==2.32.4
rich==14.0.0
ruff==0.11.13
safetensors==0.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.34.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.3
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
python-multipart==0.0.6
google-generativeai==0.8.3
